Worked on PC support
After a 65% yes to people saying RedstoneLamp should have built in PC
support, I decided to start working on the PC protocol once again. The
ping doesn't even work yet but I'm still working on getting the
fundamentals working

+++ src/main/java/net/redstonelamp/nio/BinaryBuffer.java
@@ -140,8 +140,8 @@ public class BinaryBuffer{
      *
      * @return A single unsigned short
      */
-    public int getUnsignedShort(){
-        return bb.getShort() & 0xFFFF;
+    public short getUnsignedShort(){
+        return (short) (bb.getShort() & 0xFFFF);
     }
 
     /**
